Trabajo para aplicar los filtros, pendiente de mover algunos metodos al presenter.

+++ AndroidProject/app/src/main/java/es/unican/carchargers/model/Charger.java
@@ -33,7 +33,7 @@ public class Charger {
     }
 
     // Devuelve true si alguno de los cargadores en este punto tiene la potencia indicada.
-    public boolean contienePotencia(int potencia) {
+    public boolean contienePotencia(double potencia) {
 
         //Comprobar si alguno de sus valores Power (dentro de sus connections es el indicado)
         for (Connection c : connections) {
